Beta V0.6 Improved multi-replica stability
Several changes were made to improve stability for multi-replica deployments. This many involves splitting the one deployment into a master and worker deployment.
Deployments with one replica will use the master deployment with one replica pod. Deployments with more than one replica will have a master deployment with one pod and a worker deployment with TOTALREPLICAS-1 pods. Along with it is an updated configmap deployment scheme to allow for seperate master and worker configmaps. This should help avoid the multiple masters error seen on previous version.
The graylog-auto script was updated to support this new scheme and will now dynamical apply the configmaps and deployments based on a provided manifest template.